MAT-5191 refactor measure version - version as common object
