+++ src/main/java/gov/cms/madie/models/common/ActionType.java
@@ -4,6 +4,7 @@ public enum ActionType {
   CREATED,
   VERSIONED_MAJOR,
   VERSIONED_MINOR,
+  VERSIONED_REVISIONNUMBER,
   DELETED,
   IMPORTED,
   UPDATED
