+++ src/main/java/gov/cms/madie/models/library/CqlLibrary.java
@@ -6,6 +6,7 @@ import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 import gov.cms.madie.models.validators.EnumValidator;
 import gov.cms.madie.models.utils.VersionJsonSerializer;
 import gov.cms.madie.models.common.ModelType;
+import gov.cms.madie.models.common.Version;
 
 import java.time.Instant;
 import javax.validation.GroupSequence;
